he arreclat el menu principal

+++ src/Haversine.java
@@ -4,9 +4,9 @@ public class Haversine {
     public Haversine() {
     }
 
-    public void calcularDistancia(float latitud_actual, float longitud_actual, float latitudg, float longitudg){
-        float diflat = latitudg - latitud_actual;
-        float diflong = longitudg - longitud_actual;
+    public void calcularDistancia(double latitud_actual, double longitud_actual, double latitudg, double longitudg){
+        double diflat = latitudg - latitud_actual;
+        double diflong = longitudg - longitud_actual;
 
         double a = Math.pow(Math.sin(diflat/2),2) + Math.cos(latitud_actual)*Math.cos(latitudg)*Math.pow(Math.sin(diflong/2),2);
         double c = 2*Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
